+++ application/views/main/form.php
@@ -47,7 +47,7 @@
                     </div>
                   </div>
                   <div class="pure-u-11-12">
-                    <input type="text" value="<?php echo stripslashes($i->image); ?>" name="image[]" class="pure-input-1" placeholder="Image Link">
+                    <input type="text" value="<?php echo stripslashes($i->image); ?>" name="image[<?php echo $i->imageId; ?>]" class="pure-input-1" placeholder="Image Link">
                   </div>
                   <div class="pure-u-1-24 center-text top-padding-xsmall"><a class="deleteLink" href="javascript:void(0);"><i class="fa fa-trash-o fa-7-5"></i></a></div>
                 </div>
@@ -65,7 +65,7 @@
               <?php foreach($resources_output as $r): ?>
                 <div class="pure-g">
                   <div class="pure-u-23-24">
-                    <input type="text" value="<?php echo stripslashes($r->resource); ?>" name="resource[]" class="pure-input-1" placeholder="Resource Link">
+                    <input type="text" value="<?php echo stripslashes($r->resource); ?>" name="resource[<?php echo $r->resourceId; ?>]" class="pure-input-1" placeholder="Resource Link">
                   </div>
                   <div class="pure-u-1-24 center-text top-padding-xsmall"><a class="deleteLink" href="javascript:void(0);"><i class="fa fa-trash-o fa-7-5"></i></a></div>
                 </div>